Raise file descriptor limits for broker services
Kafka Brokers want a *lot* of open files. Confluent recommends setting
the limit to at least 100,000, but the default is a paltry 4096.

We can raise this by setting LimitNOFILE in the systemd unit files for
broker services. This corresponds to setting 'ulimit -n', as per 'man
systemd.exec'.

Mirrormaker and ZooKeeper get 10,000 files, which should be plenty.
Kafka gets no limit because we are OK with it being greedy.